+++ docs/scripts/views/datepicker.vue
@@ -5,7 +5,7 @@
     </div>
 
     <div class="snippet-demo">
-      <ui-datepicker label="Select Date.." :model="date" @input="onInput" plus>
+      <ui-datepicker placeholder="Select Date.." :model="date" @input="onInput" plus>
         <i slot="plus" class="fa fa-calendar"></i>
       </ui-datepicker>
     </div>