+++ src/components/layout/tabs/panel.vue
@@ -15,33 +15,21 @@ export default {
   },
   data() {
     return {
-      currentActive: 0,
-      currentIndex: -1,
-      isActive: false
+      currentIndex: -1
     };
   },
   computed: {
     className() {
       return {
         'mdl-tabs__panel': true,
-        'is-active': this.isActive
+        'is-active': this.$parent.isActive(this.currentIndex)
       };
     }
   },
-  watch: {
-    currentActive(val) {
-      this.isActive = this.currentIndex === val;
-    }
-  },
-  updated() {
-    this.currentActive = this.$parent.currentActive;
-
-    let currentTab = this.$parent.tabs.find(value => value.name === this.tab);
+  beforeUpdate() {
+    let currentTab = this.$parent.tabs.find(tab => tab.name === this.tab);
     if (currentTab) {
       this.currentIndex = currentTab.index;
-      if (this.$parent.isActive(this.currentIndex)) {
-        this.isActive = true;
-      }
     }
   }
 };